Fix a bug related to an undefined state.results parameter
Some conditions existed comparing the state.results parameter with null, but this parameter can also be "undefined". So, when one tried to search with the searchbox empty an error occurred because it tried to access the length property of the undefined parameter. Using the logical not is possible to handle null or undefined parameters