Adds a new parameter to the advanced panel (blame URLs), updates the bindata.go file with the latest UI changes
A new parameter has been added to the advanced search panel (blame URLs), if this parameter is checked the parameter blame-url will be used instead the base-url one inside url-pattern in config.json. This is useful if one wants to go to the blame section of a file instead of go to the blob one.